+++ sga/tests/backend_test.py
@@ -1,11 +1,17 @@
 """
 Test backend functions
 """
+from io import BytesIO
+from zipfile import is_zipfile
+
 from django.core.exceptions import ValidationError
+from django.test import override_settings
+from mock import MagicMock, patch
 
 from sga.backend.authentication import get_role
 from sga.backend.constants import VALID_FILE_UPLOAD_EXTENSIONS, Roles
-from sga.backend.files import convert_illegal_S3_chars
+from sga.backend.files import convert_illegal_S3_chars, submissions_zip_generator
+from sga.backend.send_grades import send_grade, SendGradeFailure
 from sga.backend.validators import validate_file_extension
 from sga.tests.common import SGATestCase
 
@@ -52,3 +58,33 @@ class TestBackend(SGATestCase):
         }
         for unconverted, converted in CONVERSIONS.items():
             self.assertEqual(convert_illegal_S3_chars(unconverted), converted)
+
+    @patch(
+        "oauth2.Client.request",
+        MagicMock(return_value=("", "<imsx_codeMajor>success</imsx_codeMajor>"))
+    )
+    @override_settings(LTI_OAUTH_CREDENTIALS={"key": "secret"})
+    def test_send_grade(self):  # pylint: disable=no-self-use
+        """
+        Tests the send_grade() function
+        """
+        self.assertRaises(SendGradeFailure, send_grade, "not_key", "url", "result_id", 1)
+        # Call this to assert that it does NOT raise Exception
+        send_grade("key", "url", "result_id", 1)
+
+    @patch("oauth2.Client.request", MagicMock(return_value=(MagicMock(status="Bad"), "")))
+    @override_settings(LTI_OAUTH_CREDENTIALS={"key": "secret"})
+    def test_bad_send_grade(self):  # pylint: disable=no-self-use
+        """
+        Tests the failure of send_grade() function
+        """
+        self.assertRaises(SendGradeFailure, send_grade, "key", "url", "result_id", None)
+
+    def test_submissions_zip_generator(self):
+        """
+        Tests submissions_zip_generator()
+        """
+        submissions = [MagicMock(student_document=self.get_test_file()) for _ in range(10)]
+        # Since we're getting a stream, unpack streamed response
+        zipfile = bytearray("", encoding="utf8").join(submissions_zip_generator(submissions))
+        self.assertTrue(is_zipfile(BytesIO(zipfile)))
